pyproject.toml
[project]
name = "reactive-resume-auto-composer"
version = "0.1.0"
requires-python = ">=3.11"
description = "CLI tool to compose Reactive Resume JSON + Cover Letters"
readme = "README.md"
dependencies = [
  "typer[all]>=0.12.3",
  "pyyaml>=6.0.1",
  "pydantic>=2.7",
  "jinja2>=3.1",
  "rapidfuzz>=3.9",
  "orjson>=3.10"
]

[project.optional-dependencies]
web = ["httpx>=0.27", "trafilatura>=1.7", "readability-lxml>=0.8"]
pdf = ["pdfminer.six>=20240706", "pdfplumber>=0.11.4"]
docx = ["python-docx>=1.1.2"]
llm = []
dev = ["pytest", "pytest-cov", "ruff", "black", "mypy"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
